# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Planning Copilot API - AI Chat-Based Construction Scheduling Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/versions:
    get:
      operationId: getScheduleVersions
      tags: [versions]
      summary: Get schedule version history
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleVersion"
    post:
      operationId: createScheduleVersion
      tags: [versions]
      summary: Snapshot current schedule as a version
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVersionRequest"
      responses:
        "201":
          description: Created version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleVersion"
components:
  schemas:
...
    ScheduleVersion:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        versionNumber:
          type: integer
        taskCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          nullable: true
      required:
        - id
        - projectId
        - name
        - versionNumber
        - taskCount
        - createdAt
    CreateVersionRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name